// packages/sdk/src/gen/outbox/v1/channel_pb.d.ts
export declare type ListChannelsRequest = Message<"outbox.v1.ListChannelsRequest"> & {
  /**
   * Maximum number of channels to return (0 = server default, max 1000).
   *
   * @generated from field: int32 page_size = 1;
   */
  pageSize: number;

  /**
   * @generated from field: string page_token = 2;
   */
  pageToken: string;
};

/**
 * Describes the message outbox.v1.ListChannelsRequest.
 * Use `create(ListChannelsRequestSchema)` to create a new message.
 */
export declare const ListChannelsRequestSchema: GenMessage<ListChannelsRequest>;

/**
 * @generated from message outbox.v1.ListChannelsResponse
 */
export declare type ListChannelsResponse = Message<"outbox.v1.ListChannelsResponse"> & {
  /**
   * @generated from field: repeated outbox.v1.Channel channels = 1;
   */
  channels: Channel[];

  /**
   * @generated from field: string next_page_token = 2;
   */
  nextPageToken: string;

  /**
   * @generated from field: int32 total_size = 3;
   */
  totalSize: number;
};

/**
 * Describes the message outbox.v1.ListChannelsResponse.
 * Use `create(ListChannelsResponseSchema)` to create a new message.
 */
export declare const ListChannelsResponseSchema: GenMessage<ListChannelsResponse>;

/**
 * @generated from message outbox.v1.GetChannelRequest
 */
export declare type GetChannelRequest = Message<"outbox.v1.GetChannelRequest"> & {
  /**
   * Resource name. Format: channels/{channel_type}
   *
   * @generated from field: string name = 1;
   */
  name: string;
};

/**
 * Describes the message outbox.v1.GetChannelRequest.
 * Use `create(GetChannelRequestSchema)` to create a new message.
 */
export declare const GetChannelRequestSchema: GenMessage<GetChannelRequest>;

/**
 * @generated from message outbox.v1.GetChannelResponse
 */
export declare type GetChannelResponse = Message<"outbox.v1.GetChannelResponse"> & {
  /**
   * @generated from field: outbox.v1.Channel channel = 1;
   */
  channel?: Channel;
};

/**
 * Describes the message outbox.v1.GetChannelResponse.
 * Use `create(GetChannelResponseSchema)` to create a new message.
 */
export declare const GetChannelResponseSchema: GenMessage<GetChannelResponse>;

/**
 * @generated from service outbox.v1.ChannelService
 */
export declare const ChannelService: GenService<{
  /**
   * @generated from rpc outbox.v1.ChannelService.GetChannel
   */
  getChannel: {
    methodKind: "unary";
    input: typeof GetChannelRequestSchema;
    output: typeof GetChannelResponseSchema;
  },
  /**
   * @generated from rpc outbox.v1.ChannelService.ListChannels
   */
  listChannels: {
    methodKind: "unary";
    input: typeof ListChannelsRequestSchema;
    output: typeof ListChannelsResponseSchema;
  },
}>;